When Toc lines break on narrow screens, indent the same as text not open close arrow
Was going back to open close arrow, which made it confusing what was in
which level.

Also don't show word counts on ToC by default. Too much noise. On hover/click
only.

Also separate toc items a little bit more to make it easier to
differentiate different entries when there is no number and lines are
wrapping in narrow screens.