Proper pagination and tabs with separate URLs
Just noticed that because of JWT, we simply cannot do SSR for logged-in
pages. For now just using API for the requests that require login such as
top-followed, but what we really have to do is to migrate
to a proper Cookie authentication, and then just do everything from Next.js
without any API GETs.

Topics also working now, and added latest.