To implement instead of doing DELETE we just do a AND defined_at != current path for every query.
This led to realizing once again that INCLUDE was not well factored
out with the absolutely identical X and X_CHILD mechanisms that came later
and so we unify them to make the implementation sane.

Also fix a newly noticed bug that tag= was not having an effect on Tagged
across files, one of those rare cases for Ciro where the bug is found by
reading the source theoretically rather than in an experiment.