Get rid of the mock provider, use sqlite memory
This significantly slows down a run of:

npm test -- -g '^(?!executable: )'

from 1.1s to 3s, but I think it is worth it, as it will allow moving more
tests to non-filesystem mode, and executable tests are the ones that are
really slow.

Also it saves the development effort of having to update the mock whenever
there is a change in IdProvider. And it tests the provider more. Let's go.