Make multiple include parents show again and test that
This means generalizing get_header_parent to get_header_parents.

Every point of this patch that adds a [0] as a quick hack, which is
every point except header render, is a possible missing feature, we didn't
fully check the implications of each one of them, it's just not
worth the time right now.

One that is definitely a bug is SqliteIdProvider.update and the underlying
data structure one to many, to support multiple parents we need many to
many there and therefore another table.