Convert include refs to parent refs
We now store every single indexed parent to child relationship on the
database in the Refs table.

Previously only parent relationships by \Include were stored.

This is the crucial step for cross file ToCs. Now we "just" have to query
all the children recursively out and put it onto the ToC. Easy right?
But the data is clean on the DB now for us to access.