Fix link to image in another file that has a title with link to a third file
This change makes \x not render anymore for ID determination, href is used
as is. This is the only way to be able to calculate all IDs in a single
ID extraction pass, otherwise we would need yet another ID, and the result
is not much better, so not doing it.

This more clearly suggests the principle that ID extraction should never
rely on the database state itself:
https://github.com/cirosantilli/cirodown/issues/208 only on one single
input file at a time.

As a first requisite of that principle, we move here the title to title
checks after Id extraction in cirodown, much like we had done for ID
duplicate checks. Unfortunately we lose the exact line/column information
of the \x culprit, and retain only the source location of the image itself.
TODO improve later.

Also:

- only check duplicate IDs for the files being currently converted.
  Otherwise it would be confusing for users to get errors on files that
  don't have anything to do with the current one being converted.
- don't print extract_id twice with `cirodown --no-render .` A false
  one was being printed for the initial file gathering pass.
- don't add an empty metadata line on the toplevel header.
  An undefined vs '' bug.