Fix relative include in subdirectories
This became fundamental because we are scoping every dynamic website
article by the author ID @username.

Fix https://github.com/cirosantilli/cirodown/issues/161

To do this, we moved the goalposts very slightly by making
the h1 of subdir/index.ciro have no id. This is more consistent
with the general procedure of scope removal, and we never want to link to
that header anyways. Not removing it would also potentially lead to double
IDs on the file due to the removal inconsistency.