Convert subdir/index.ciro to subdir.html instead of subdir/index.html
The main motivation for this is to not break index links on the dynamic
website, which does not have trailing slashes. Trailing slashes are added
by GitHub in the case of subdir/index.html but not subdir.html.

This is also feels slightly more consistent, same toplevel ID goes to
the same output path always.