Revive ancestors, make work cross files, and test it this time
Best commit message every, it was an accident I swear.

To make it work across files we use another RECURSIVE query that walks up
the parent asts until the top.

Note that cross parent already seemed to work on the tests/cirodown .
directory conversion due to caching. But if you pick just a single file
in cirodown input.ciro it wasn't working. Not it is.

Fix https://github.com/cirosantilli/cirodown/issues/219