web: get rid of LikeArticleButtonContext
I'm just not smart enough to handle this. It is something that came from
Thinkster, in theory would be nice to have multiple buttons on a single
page turn color. But we don't have any in this website, and I'm just not
smart enough to make this work, notably the . So be it, I can live with that.