cli: don't do DB checks if there was a conversion error
When splitting large files, this could lead to hundreds of false duplicates
if there was an error on the new file, making it hard to find the actual
error. E.g.:

b.ciro

```
= b

== a
```

then convert `cirodown .` and modify to:

a.ciro

```
= a
```

b.ciro

```
= b
```

and reconvert. This would complain about a duplicate ID a.

At https://github.com/cirosantilli/cirodown/tree/gather-timestamps we also
solved the problem by gathering all timestamp checks and deletions in one
place. That could be potentially slightly more efficient as it could do
less DB accesses, but actually making things more efficient would be a bit
complicated, as we would need to remove the individual deletions that we
are currently doing and readd them to several places. Embed includes also
poses further complications, as we need to first parse before we can know
the full list of what has to be deleted in that case. So we just KISSed it
for now.