web: make it possible to test Next.js
For this to work on SQLite, we stop reconnecting to the database again from
Next.js in addition to the connection made from app.js at startup. This is
necessary because there is no way to access the same in-memory database
with two separate connections. It is also much generally sane anyways.