web: cli: OK, fix upload by explicitly giving path everytime
It's just the simplest thing to do for now... not ideal because not
possible to do id= from body... but so be it. We could enable that with
a separate dedicated field... which is what we will eventually have to do
anyways on the final GUI setup... so be it.