web: redirect user to register page from article editor if not logged in
We do have some links that show up to new article when the user is not
logged in.

Those are left in places where we don't have enough space to ask "Sign up
or sign in" in full. But we like to keep those "Create" links to make it
more likely that people will sign up.

If I break Navbar display again, this commit is the culprit.