web: implement \H toplevel argument
Remove the \Toc macro entirely. This was something we've wanted to do for a
while, as we don't want variation in Toc positions on Web. And it came up
naturally as part of the refactor for toplevel, so it is done. It is now
impossible to customize ToC location. Optinionation for the win.
The Toc is now render-time only, it does not show on the Ast at all, which
is much cleaner.

Toc position testing is now done with instrumentation on HTML output.

Getting this patch to work without breaking everything else made me so
angry I got a stomachache for half a day. I shouldn't do this, but I can't
control myself... a big part of the problem was that I also refactored a
billion things in addition to the minimal patch. Then it was impossible to
tell where error came from. I had to revert the minimal patch, get old things
working again and then reapply and fix just the patch. All in a very very
messy way.