web: fetch all local article shas at once locally
Previously we were fetching them one by one. Now we fetch them in bulk.
This reduces the time to first render from about 5s to zero.

We then only fetch the data for articles that we are actually going to
render. This speeds up differential rendering, as we skip fetching the
majority of files that we don't care about one by one.

Also improve the Cli --web username default handling. Create the concept
of defaultUsernameForHost, and use the last login by default for now.