web: fix all issues giving 500
The error was:

TypeError: Cannot read properties of undefined (reading 'split')

and it was introduced when we started having dynamic ancestor lists on web.

Also remove ancestor list div from issues, as that would be the next
blowup.

Create rerender-issues, analogous to rerender-articles.

Add --author option to rerender-articles to select only one author.