package.json: update node engine to 20.10.0
This also required updating webpack which was incompatible with the older node.

Which in turn also required updating Next.js to something with the newer
webpack, so we attempted the latest v14. Arghh. OK, it wasn't so painful
after all, just two errors and one of them covered by the codemod.