web: prevent /go/user/[uid]/articles click to /go/user/[uid]/discussions from showing discussion bodies
This was happening because the ShowBody button is not present in
discussions, and thus failed to reset the body state.

The rule is that useState resets must always be together with the useState
call. Maybe we should have created a custom hook to factor things in this
case, but lazy.

Also fix double left margin on /go/user/[uid]/* user lists.