web: fix 500 if unlisted article that is first of its depth has listed child
The 500 happens when visiting the parent of the child on Index rendering.

This was breaking https://ourbigbook.com/pioyi and was reported at:
https://ourbigbook.com/go/discussion/2/pioyi/oveview#_comment-3

We were already previously handling the case of unlisted article with listed child
by reparenting it on the ToC to the first listed ancestor, but this edge case
was missed.

Also fix toc construction in closely related edge cases, now more correctly
reparenting nodes.