cli: prevent double include of a single file
In the past I had some pipe dreams of allowing multiple includes.

But it is insane and has no use case and violates the sanctity of the
header tree and is not web uploadable and leads to several edge cases.
So let's nuke that now once and for all.

Fix https://github.com/ourbigbook/ourbigbook/issues/223

Fix https://github.com/ourbigbook/ourbigbook/issues/189