cli: prevent files without a parent include
This would leave out files that are not shown on the static website toc,
and most importantly is not web uploadable.

Fix https://github.com/ourbigbook/ourbigbook/issues/269

Fixing the tests that broke due to this commit almost killed me. But I
survived.

cli: don't delete all IDs outside of a subdirectory when converting just the subdirectory