web: show tagged headers under non-toplevel headers
Same as for static on the previous commit, but now on web.

Doing a dynamic implementation for now. Let's see if it hits perf. If it
does, then at one point we have to just start caching a pre-rendered links
list per each article and updating that on each tag change.