web: make ourbigbook --web make unlisted articles listed even if content didn't change
Before this change, if an article was made unlisted, either on web UI or CLI,
and then later on you try to do --web upload with the same content, the
article would not get updated and would remain unlisted.

Now it will become listed. This was implemented by adding list to the hash
where it belongs.

Also: prevent the index page from being unlisted via API... That would be
bad.