web: keep GET "search" param exactly as in text field
Previously we were doing title to ID on it and storing that on the URL.

But that is not very good because then if you copy and paste the URL
e.g. with search "test data", you search is now "test-data".

Rather, we must store the search exactly as it is on the text area,
and then use that to fill the search bar on UI, and let the server
do titleToId conversion, which this commit does.