web: fix Firefox frontend blowup due to new URL throwing on invalid URLs
The issue affected both:
* http://localhost:3000/barack-obama/test-data
* http://localhost:3000/go/articles?search=tes

for an input of type:
``
http://example.com\u{00A0}asdf
``
where 00A0 is a literal &nbsp;

Blows up on Firefox 135 but not on Chromium 132, Ubuntu 24.10. Firefox
appears to be more strict in its new URL parsing which is annoying.

The error was first exposed by
https://ourbigbook.com/margaretschiel43/recover-your-funds-with-confidence-through-pyramid-hack-solution
due to some randomly formatted spammer input: https://github.com/ourbigbook/ourbigbook/issues/346