web: fix incorrect infinite parent cycle check based on nested set index when bulk updating
We need to do cycle check on the Ref because nested set is not necessarily
up to date at each step.

This fixes the issue reported on the following comment:
https://github.com/ourbigbook/ourbigbook/issues/319#issuecomment-2662912799

It is unclear if the original problem of the issue is solved or not because
it is hard to reproduce.