lib: bigb output: keep scope exactly as it is on input source
Previously we were lowercasing it and removing characters that don't
contribute to ID. But that makes it more likely that the scope will
become less readable.

Ideally what we really want is to get all scope Ids on a recursive query
and render them like the last component. But lazy to implement that now,
and this already improves the output a lot.