lib: bigb output: try to get dollars in <> and fail for now
Added a failing test and summarized status at:
https://github.com/ourbigbook/ourbigbook/issues/362

It's not worth to work on this further for now, was just a quick try
in case it was easy. It's good enough without for now.

Also fix

``
= \testSaneOnly[a]
``

giving insane ID because it was converting a function object to a string.
I love js.