lib: bigb output: move id to fixed order as second element after title
IDs are too important for similar reasons as the header. It would be bad
if the ID were to fall after a huge block like a multi paragraph
description.

With other short args, it's not so bad however and alphabetic is slightly
better perhaps:

``
= Asdf
{c}
{id=qwer}
{title2=Asdf}
``

But I think ID on top is better generally.