lib: fix SqlDbProvider.fetch_header_tree_ids definedAtFileId to not return includes in other files
The bug was visible when given:

index.bigb

``
= Toplevel

\Include[notindex]

== Toplevel 2
``

notindex.bigb

``
= Notindex

== Notindex 2
``

and you convert just one file:

``
ourbigbook --web index.bigb
``

was also converting Notindex.

The issue was that Ref to_id was being used and the defined at check,
was done on Ref, and the Ref defined in index.bigb caught Notindex
in the other file.