web: consider web conversion as index if the path is 'index'
Previously the path was ignored. Now it is considered.

This fixes rerender-articles for index articles that had a non-empty
titleSource which was the case for cirosantilli2 for example on
ourbigbook.com. This did not seem possible to achieve just before this
commit, but it must have been possible in the past, and will become
possible again from now on.

Also don't automatically make the index titleSource empty when editing
the index page on web editor, keep whatever is on database as it is, even
though it doesn't show up in any way. This could be used for source
download later on, as it stores information that is present locally.

Fix https://github.com/ourbigbook/ourbigbook/issues/357