web: show scope on web indices
A perhaps more ideal solution would be to store the fully scoped titles
on Id so we can then also use that on static render incoming links and
tagged. But that is slightly harder to implement as it requires a new pass
phase that would calculate the scoped titles, since scopes in general
depend on all ancestors. We could simplify things a bit if we set ourselves
to a single file, and {scope} only works in a single file as a limitation,
but directory based scope already works across files so it wouldn't work
there...

Another alternative for static would be to have a new Ref that points to
all scoped ancestors of each ID, then we could easily fetch everything
needed and efficiently render it at rendertime with the correct number of
scopes resolved. For another day I think.

Contributes towards https://github.com/ourbigbook/ourbigbook/issues/144