cli: don't show every single conversion as skipped by timestamp
For bigb files, the issue was just a flag being set incorrectly.

For non-bigb files, the issue was an unintended pre-conversion.

Funny how this stayed broken like this for such a long time and I never
noticed, only to spot it while trying to reproduce another bug from the
cli. Funny and sad.