web: move article author from above title to breadcrumb below
I've wanted to move author down for a long time, but now I've found a
perfect location for it: at the start of the breadcrumb, replacing what
previously showed as "Home".

This both shows prominently as the first thing below the title as desired,
and also intuitively fits in with the pre-existing ancestor breadcrumb.
Beauty.